fix SR edit form
